Fix for inconsistent JSObject impl. in JVM's
In some scenario's different impl. of
netscape.javascript.JSOBject.call(...) is used. 

Inside Developer the version that is used uses Object[] as second
argument, whereas when deployed through WebStart an vararg is used as
second parameter. 

This fix tests which version is available and branches code based on the
result of the test.